Feature/astparser update

+++ modules/astparser/src/ast/declarations/ASTMemberGroupDeclaration.h
@@ -57,7 +57,7 @@ public:
     }
 
     virtual ~ASTMemberGroupDeclaration() {
-        for (unsigned int i = 0, count = members.size(); i < count; i++)
+        for (size_t i = 0, count = members.size(); i < count; i++)
             delete members[i];
         members.clear();
     }
