+++ modules/astparser/src/ast/expressions/ASTSwitchBodyExpression.h
@@ -51,7 +51,7 @@ public:
     ASTSwitchBodyExpression() : ASTExpression() {}
 
     virtual ~ASTSwitchBodyExpression() {
-        for (int i = 0, count = cases.size(); i < count; i++)
+        for (size_t i = 0, count = cases.size(); i < count; i++)
             delete cases[i];
         cases.clear();
     }
